# docs/swagger.yaml
swagger: "2.0"
info:
  description: "Полнофункциональный REST API для игры в покер с категориями столов LOW, MID, VIP. Использует PostgreSQL, Kafka, Redis и авторизацию через Telegram Web App"
  title: "Poker REST API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    name: "API Support"
    url: "http://www.swagger.io/support"
    email: "support@swagger.io"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
  version: "1.0"
host: "localhost:3000"
basePath: "/api/v1"
securityDefinitions:
  TelegramAuth:
    type: apiKey
    name: x-init-data
    in: header
    description: "Telegram Web App init_data для авторизации"
paths:
...
  /profile:
    get:
      tags:
        - users
      summary: "Получить профиль пользователя"
      description: "Возвращает профиль авторизованного пользователя"
      security:
        - TelegramAuth: []
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        200:
          description: "OK"
          schema:
            type: object
            properties:
              user:
                $ref: "#/definitions/User"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
    put:
      tags:
        - users
      summary: "Обновить профиль пользователя"
      description: "Обновляет данные профиля авторизованного пользователя"
      security:
        - TelegramAuth: []
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: profile
          in: body
          required: true
          description: "Данные для обновления"
          schema:
            type: object
            properties:
              username:
                type: string
      responses:
        200:
          description: "OK"
          schema:
            type: object
            properties:
              message:
                type: string
              user:
                $ref: "#/definitions/User"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  User:
    type: object
    properties:
      uuid:
        type: string
      username:
        type: string
      telegram_id:
        type: integer
      balance:
        type: integer
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
...
  Error:
    type: object
    properties:
      error:
        type: string